polish(music): world-class spec on music-lab + catalog surfaces
